Change name and location of StaticVariable
