+++ src/decomon/layers/__init__.py
@@ -1,2 +1,3 @@
-from .core import ForwardMode, StaticVariables
+from ..core import InputsOutputsSpec
+from .core import ForwardMode
 from .utils import get_lower, get_upper, maximum, minimum
