+++ src/decomon/models/models.py
@@ -4,8 +4,13 @@ from typing import Any, Dict, List, Optional, Union
 import tensorflow as tf
 from keras.engine.functional import get_network_config
 
-from decomon.core import BoxDomain, GridDomain, Option, PerturbationDomain
-from decomon.layers.core import StaticVariables
+from decomon.core import (
+    BoxDomain,
+    GridDomain,
+    InputsOutputsSpec,
+    Option,
+    PerturbationDomain,
+)
 from decomon.models.utils import ConvertMethod
 
 
@@ -28,7 +33,7 @@ class DecomonModel(tf.keras.Model):
         if perturbation_domain is None:
             perturbation_domain = BoxDomain()
         self.perturbation_domain = perturbation_domain
-        self.nb_tensors = StaticVariables(dc_decomp).nb_tensors
+        self.nb_tensors = InputsOutputsSpec(dc_decomp).nb_tensors
         self.dc_decomp = dc_decomp
         self.method = ConvertMethod(method)
         self.ibp = ibp
